% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_datasets.R
\name{ndap_list_datasets}
\alias{ndap_list_datasets}
\title{List All Available NDAP Datasets}
\usage{
ndap_list_datasets()
}
\value{
A tibble (data frame) containing dataset metadata,
including ID, title, organization, and update frequency.
}
\description{
Connects to the NDAP API to retrieve a catalog of all currently
available datasets.
}
\examples{
\dontrun{
if (Sys.getenv("NDAP_KEY") != "") {
  datasets <- ndap_list_datasets()
  print(head(datasets))
}
}
}
