% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/key_management.R
\name{ndap_set_key}
\alias{ndap_set_key}
\title{Set the NDAP API Key as an Environment Variable}
\usage{
ndap_set_key(key)
}
\arguments{
\item{key}{The character string containing your NDAP API key.}
}
\value{
The function is called for its side effect of setting
the environment variable. It returns \code{NULL} invisibly.
}
\description{
This function saves your personal NDAP API key to your
\code{.Renviron} file, allowing the \code{randap} package
to access the API without hardcoding the key.
}
\examples{
\dontrun{
ndap_set_key("YOUR_ACTUAL_API_KEY_HERE")
}
}
