% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nsso_layouts.R
\docType{data}
\name{nsso_2024_layout_L01}
\alias{nsso_2024_layout_L01}
\title{HCES 2023-24 Level 01 Layout}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 21 rows and 2 columns.
}
\usage{
nsso_2024_layout_L01
}
\description{
HCES 2023-24 Level 01 Layout
}
\keyword{datasets}
