% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_mpce.R
\name{rnsso_calc_mpce}
\alias{rnsso_calc_mpce}
\title{Calculate Monthly Per Capita Expenditure (MPCE)}
\usage{
rnsso_calc_mpce(data_list)
}
\arguments{
\item{data_list}{The list output from rnsso_read_data() folder load.}
}
\description{
Calculate Monthly Per Capita Expenditure (MPCE)
}
