% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/item_codes.R
\name{rnsso_label_items}
\alias{rnsso_label_items}
\title{Label HCES Item Codes}
\usage{
rnsso_label_items(df)
}
\arguments{
\item{df}{A dataframe containing an 'item_code' column.}
}
\description{
Label HCES Item Codes
}
