% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_data.R
\name{rnsso_read_data}
\alias{rnsso_read_data}
\title{Read NSSO/HCES Data (Single File or Full Folder)}
\usage{
rnsso_read_data(path, round = 2024, level = NULL, delim = "\\t")
}
\arguments{
\item{path}{Path to either a single .txt file OR a folder containing multiple LEVEL files.}

\item{round}{The year/round of the survey (default 2024).}

\item{level}{Only required if reading a single file (e.g., "L01").}

\item{delim}{The delimiter used (default is tab "\t").}
}
\description{
Read NSSO/HCES Data (Single File or Full Folder)
}
